python/requirements.txt
numpy
matplotlib
opencv-python
Pillow
pywin32
pynput